NMSBackend/requirements.txt
Flask==2.3.2
firebase-admin==6.0.1
google-cloud-firestore==3.12.1
